services/conversion/requirements.txt
fastapi[standard]
faststream[nats]
conductor-python